# PERF: ArrowExtensionArray._from_sequence